include title and statement pages

+++ docs/generate.fsx
@@ -34,10 +34,11 @@ let createPDF fileName =
     p.StartInfo.RedirectStandardOutput <- false
     p.Start() |> ignore
     p.WaitForExit()
+    if p.ExitCode <> 0 then failwith "pdflatex failed"
     for ext in ["tex"; "aux"; "out"; "log"] do
         let auxFile = Path.ChangeExtension(fileName, ext)
         printfn "Delete auxiliary file: %s" auxFile
-        File.Delete(auxFile)
+        if File.Exists(auxFile) then File.Delete(auxFile)
     Path.ChangeExtension(fileName, "pdf")
 
 CreateDir outputDir
